refactored parser
Made a lot of the parse methods private because I really only want users
to be able to parse either an entire document or a template body. I had
to add another method called parse_template that allows you to parse the
body of a template without the <template></template> tags. That parse
method still returns a Template AST node but the node will not have any
attributes and the location will be the beginning and end of the
children instead of the start and close tags.