fixed lexer state issue and tests
The initial scan shouldn't happen in the initialize method of the parser
because we need to make sure the lexer is in the right state before
doing the first scan. Since the user can call the parse_template method
or the parse method it's not clear at time of initialization what the
state should be. So moved the first scan into the two public methods:
parse and parse_template. This also required updates to the parser tests
to do the initial scan manually for the private method tests.