added document ast require statement
We need to require "elements/template" at the top of a document so that
the Template class is available in the compiled file. Also removed
auto creating a newline before each document node. Instead, the document
node itself will determine whether to create a newline (so I don't
create redundant newlines)