removed active_support dependency
I don't want to introduce unnecessary dependencies, especially on Rails
packages because this increasees the probability of a version conflict
in projects. For example, if you're using a previous version of Rails
and I'm requiring the latest version we might have a problem. So I'll
put any utility methods I need in the elements/utils.rb file and those
utilities can be included as a module.